requirements.txt
streamlit>=1.35.0
groq>=0.9.0
opencv-python-headless>=4.9.0
Pillow>=10.3.0
rembg[cpu]>=2.0.50
mediapipe>=0.10.0
numpy>=1.26.0
requests>=2.31.0
